+++ restcrab/src/crabs/reqwest.rs
@@ -83,6 +83,9 @@ impl crate::Restcrab for Reqwest {
     for (key, value) in &request.headers {
       req_builder = req_builder.header(key, value);
     }
+
+    req_builder = req_builder.query(&request.queries.iter().collect::<Vec<_>>());
+    
     if let Some(body) = &request.body {
       req_builder = req_builder.body(serde_json::to_string(body).context(SerializingBodySnafu)?);
     }