+++ restcrab_macros/src/expansion/on_trait.rs
@@ -93,7 +93,7 @@ pub fn on_trait(args: &super::Args, input: &mut syn::ItemTrait) -> Result<TokenS
         if let syn::FnArg::Typed(pat_type) = parameter {
           pat_type
             .attrs
-            .retain(|a| a.path != syn::Path::from_string("body").unwrap() && a.path != syn::Path::from_string("headers").unwrap());
+            .retain(|a| a.path != syn::Path::from_string("body").unwrap() && a.path != syn::Path::from_string("headers").unwrap() && a.path != syn::Path::from_string("queries").unwrap());
         }
       }
     }