+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add support for query parameters
+
 ## [0.1.2] - 2022-01-25
 ### Changed
 - Update dependencies